requirements.txt
aiohttp 
aiosignal 
anyio==3.7.0
async-timeout 
attrs 
backoff==2.2.1
certifi==2023.5.7
charset-normalizer 
chromadb==0.3.25
click 
clickhouse-connect==0.5.25
colorama 
coloredlogs==15.0.1
dataclasses-json 
duckdb==0.8.0
fastapi==0.96.0
ffmpeg-python==0.2.0
filelock==3.12.0
filetype==1.2.0
Flask 
Flask-Cors==3.0.10
flatbuffers==23.5.26
frozenlist 
future==0.18.3
greenlet 
Gunicorn
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
humanfriendly==10.0
idna 
itsdangerous 
Jinja2 
langchain 
llvmlite==0.40.0
lz4==4.3.2
MarkupSafe
marshmallow 
marshmallow-enum 
# mkl-fft
# mkl-random 
# mkl-service==2.4.0
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict 
mypy-extensions 
networkx==3.1
numba==0.57.0
numexpr 
numpy 
onnxruntime==1.15.0
openai==0.27.7
openai-whisper==20230314
openapi-schema-pydantic 
overrides==7.3.1
packaging
posthog==3.0.1
protobuf==4.23.2
pydantic 
pydub==0.25.1
pypdf==3.9.0
PyPDF2 
pyreadline3==3.4.1
PySocks 
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML 
regex==2023.5.5
requests 
six==1.16.0
sniffio==1.3.0
SQLAlchemy 
starlette==0.27.0
stringcase==1.2.0
sympy==1.12
tenacity 
tiktoken==0.3.1
tokenizers==0.13.3
torch==2.0.1
tqdm 
typing-inspect 
typing_extensions 
tzdata==2023.3
urllib3 
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug 
whisper==1.1.10
win-inet-pton 
yarl 
zstandard==0.21.0
